src/test/java/com/blogapp/security/UserDetailsServiceImplTest.java
package com.blogapp.security;

import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

import java.util.Optional;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;

import com.blogapp.user.entity.User;
import com.blogapp.user.repository.UserRepository;

@ExtendWith(MockitoExtension.class)
public class UserDetailsServiceImplTest {

    @Mock
    private UserRepository userRepository;

    @InjectMocks
    private UserDetailsServiceImpl userDetailsServiceImpl;

    @Test
    void whenUserNotFoundThrowsBadCredentialsException() {
        when(userRepository.findByUsername(anyString())).thenReturn(Optional.empty());

        assertThrows(UsernameNotFoundException.class, () -> {
            userDetailsServiceImpl.loadUserByUsername("username");
        });
    }

    @Test
    void whenUserFoundReturnsUserDetails() {
        when(userRepository.findByUsername(anyString())).thenReturn(Optional.of(mock(User.class)));

        var result = userDetailsServiceImpl.loadUserByUsername("username");
        assertTrue(result instanceof UserDetails);
    }
}
